requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.20.0
transformers>=4.30.0
opencv-python>=4.7.0
numpy>=1.24.0
scipy>=1.10.0
Pillow>=9.5.0
safetensors>=0.3.0
openai>=0.27.0
tqdm>=4.65.0
accelerate>=0.20.0
python-dotenv>=1.0.0
